# split response in bucket series