Benchmark result:

- benchstat

```
name                                                     old time/op    new time/op    delta
Series/10e6SeriesWithOneSample/1of10000000-40               4.83s ± 2%     4.90s ± 2%    ~     (p=0.222 n=5+5)
Series/10e6SeriesWithOneSample/10of10000000-40              4.86s ± 4%     4.92s ± 6%    ~     (p=0.421 n=5+5)
Series/10e6SeriesWithOneSample/100of10000000-40             4.95s ± 5%     5.07s ± 9%    ~     (p=0.151 n=5+5)
Series/10e6SeriesWithOneSample/1000of10000000-40            5.01s ± 5%     4.91s ± 1%    ~     (p=0.690 n=5+5)
Series/10e6SeriesWithOneSample/10000of10000000-40           4.99s ± 6%     4.98s ± 1%    ~     (p=0.690 n=5+5)
Series/10e6SeriesWithOneSample/100000of10000000-40          4.88s ± 2%     5.08s ± 4%  +4.27%  (p=0.008 n=5+5)
Series/10e6SeriesWithOneSample/1000000of10000000-40         6.54s ± 8%     6.40s ± 3%    ~     (p=0.548 n=5+5)
Series/OneSeriesWith100e6Samples/1of100000000-40           36.5ms ± 2%    37.5ms ± 1%  +2.74%  (p=0.008 n=5+5)
Series/OneSeriesWith100e6Samples/10of100000000-40          36.8ms ± 7%    37.5ms ± 1%    ~     (p=0.690 n=5+5)
Series/OneSeriesWith100e6Samples/100of100000000-40         36.2ms ± 5%    35.9ms ± 3%    ~     (p=0.841 n=5+5)
Series/OneSeriesWith100e6Samples/1000of100000000-40        36.3ms ± 4%    36.5ms ± 1%    ~     (p=0.841 n=5+5)
Series/OneSeriesWith100e6Samples/10000of100000000-40       37.6ms ± 3%    37.7ms ± 3%    ~     (p=1.000 n=5+5)
Series/OneSeriesWith100e6Samples/100000of100000000-40      37.6ms ± 3%    38.0ms ± 2%    ~     (p=0.690 n=5+5)
Series/OneSeriesWith100e6Samples/1000000of100000000-40     40.2ms ± 3%    40.1ms ± 2%    ~     (p=0.690 n=5+5)
Series/OneSeriesWith100e6Samples/10000000of100000000-40     129ms ± 4%     128ms ± 4%    ~     (p=0.690 n=5+5)

name                                                     old alloc/op   new alloc/op   delta
Series/10e6SeriesWithOneSample/1of10000000-40              1.58GB ± 0%    1.58GB ± 0%    ~     (p=0.690 n=5+5)
Series/10e6SeriesWithOneSample/10of10000000-40             1.58GB ± 0%    1.58GB ± 0%    ~     (p=1.000 n=5+5)
Series/10e6SeriesWithOneSample/100of10000000-40            1.58GB ± 0%    1.58GB ± 0%    ~     (p=0.548 n=5+5)
Series/10e6SeriesWithOneSample/1000of10000000-40           1.58GB ± 0%    1.58GB ± 0%    ~     (p=0.421 n=5+5)
Series/10e6SeriesWithOneSample/10000of10000000-40          1.59GB ± 0%    1.59GB ± 0%  +0.01%  (p=0.008 n=5+5)
Series/10e6SeriesWithOneSample/100000of10000000-40         1.67GB ± 0%    1.67GB ± 0%  -0.12%  (p=0.029 n=4+4)
Series/10e6SeriesWithOneSample/1000000of10000000-40        2.54GB ± 0%    2.53GB ± 1%    ~     (p=0.310 n=5+5)
Series/OneSeriesWith100e6Samples/1of100000000-40           59.3MB ± 0%    59.3MB ± 0%    ~     (p=0.730 n=4+5)
Series/OneSeriesWith100e6Samples/10of100000000-40          59.3MB ± 0%    59.3MB ± 0%    ~     (p=0.452 n=5+5)
Series/OneSeriesWith100e6Samples/100of100000000-40         59.3MB ± 0%    59.3MB ± 0%    ~     (p=0.889 n=5+5)
Series/OneSeriesWith100e6Samples/1000of100000000-40        59.2MB ± 0%    59.2MB ± 0%    ~     (p=0.841 n=5+5)
Series/OneSeriesWith100e6Samples/10000of100000000-40       59.4MB ± 0%    59.4MB ± 0%    ~     (p=0.151 n=5+5)
Series/OneSeriesWith100e6Samples/100000of100000000-40      60.4MB ± 0%    60.4MB ± 0%    ~     (p=0.690 n=5+5)
Series/OneSeriesWith100e6Samples/1000000of100000000-40     69.1MB ± 0%    69.1MB ± 0%    ~     (p=0.310 n=5+5)
Series/OneSeriesWith100e6Samples/10000000of100000000-40     298MB ± 0%     298MB ± 0%  +0.00%  (p=0.008 n=5+5)

name                                                     old allocs/op  new allocs/op  delta
Series/10e6SeriesWithOneSample/1of10000000-40               20.1M ± 0%     20.1M ± 0%    ~     (p=0.548 n=5+5)
Series/10e6SeriesWithOneSample/10of10000000-40              20.1M ± 0%     20.1M ± 0%    ~     (p=0.841 n=5+5)
Series/10e6SeriesWithOneSample/100of10000000-40             20.1M ± 0%     20.1M ± 0%    ~     (p=0.690 n=5+5)
Series/10e6SeriesWithOneSample/1000of10000000-40            20.1M ± 0%     20.1M ± 0%    ~     (p=0.548 n=5+5)
Series/10e6SeriesWithOneSample/10000of10000000-40           20.1M ± 0%     20.1M ± 0%    ~     (p=0.841 n=5+5)
Series/10e6SeriesWithOneSample/100000of10000000-40          20.6M ± 0%     20.6M ± 0%    ~     (p=1.000 n=5+5)
Series/10e6SeriesWithOneSample/1000000of10000000-40         25.1M ± 0%     25.1M ± 0%    ~     (p=1.000 n=5+5)
Series/OneSeriesWith100e6Samples/1of100000000-40              270 ± 0%       269 ± 0%  -0.37%  (p=0.029 n=4+4)
Series/OneSeriesWith100e6Samples/10of100000000-40             270 ± 0%       269 ± 0%    ~     (p=0.095 n=4+5)
Series/OneSeriesWith100e6Samples/100of100000000-40            270 ± 0%       269 ± 0%    ~     (p=0.095 n=5+4)
Series/OneSeriesWith100e6Samples/1000of100000000-40           288 ± 0%       287 ± 0%    ~     (p=0.095 n=4+5)
Series/OneSeriesWith100e6Samples/10000of100000000-40          447 ± 0%       447 ± 0%    ~     (all equal)
Series/OneSeriesWith100e6Samples/100000of100000000-40       1.98k ± 0%     1.98k ± 0%    ~     (p=0.762 n=5+5)
Series/OneSeriesWith100e6Samples/1000000of100000000-40      17.2k ± 0%     17.2k ± 0%  +0.09%  (p=0.008 n=5+5)
Series/OneSeriesWith100e6Samples/10000000of100000000-40      169k ± 0%      170k ± 0%  +0.09%  (p=0.008 n=5+5)
```

- benchcmp

```
benchmark                                                            old ns/op      new ns/op      delta
BenchmarkSeries/10e6SeriesWithOneSample/1of10000000-40               4856638184     4946484703     +1.85%
BenchmarkSeries/10e6SeriesWithOneSample/1of10000000-40               4794844978     4885130501     +1.88%
BenchmarkSeries/10e6SeriesWithOneSample/1of10000000-40               4749861366     4808052451     +1.23%
BenchmarkSeries/10e6SeriesWithOneSample/1of10000000-40               4819983487     4854448965     +0.72%
BenchmarkSeries/10e6SeriesWithOneSample/1of10000000-40               4937215886     5002189868     +1.32%
BenchmarkSeries/10e6SeriesWithOneSample/10of10000000-40              4812740603     4905060537     +1.92%
BenchmarkSeries/10e6SeriesWithOneSample/10of10000000-40              4811120007     4802431838     -0.18%
BenchmarkSeries/10e6SeriesWithOneSample/10of10000000-40              4763646975     4813847109     +1.05%
BenchmarkSeries/10e6SeriesWithOneSample/10of10000000-40              4853549806     4864929830     +0.23%
BenchmarkSeries/10e6SeriesWithOneSample/10of10000000-40              5070036915     5196624250     +2.50%
BenchmarkSeries/10e6SeriesWithOneSample/100of10000000-40             4874739981     5526027039     +13.36%
BenchmarkSeries/10e6SeriesWithOneSample/100of10000000-40             4945885975     5010509566     +1.31%
BenchmarkSeries/10e6SeriesWithOneSample/100of10000000-40             4852541927     4906544733     +1.11%
BenchmarkSeries/10e6SeriesWithOneSample/100of10000000-40             4889605754     4950472290     +1.24%
BenchmarkSeries/10e6SeriesWithOneSample/100of10000000-40             5191323806     4963634989     -4.39%
BenchmarkSeries/10e6SeriesWithOneSample/1000of10000000-40            5219882021     4913290699     -5.87%
BenchmarkSeries/10e6SeriesWithOneSample/1000of10000000-40            5002479712     4924810732     -1.55%
BenchmarkSeries/10e6SeriesWithOneSample/1000of10000000-40            5233585428     4880437240     -6.75%
BenchmarkSeries/10e6SeriesWithOneSample/1000of10000000-40            4806390220     4839426673     +0.69%
BenchmarkSeries/10e6SeriesWithOneSample/1000of10000000-40            4765568516     4968656365     +4.26%
BenchmarkSeries/10e6SeriesWithOneSample/10000of10000000-40           5078833696     5005195206     -1.45%
BenchmarkSeries/10e6SeriesWithOneSample/10000of10000000-40           4874589683     4947345009     +1.49%
BenchmarkSeries/10e6SeriesWithOneSample/10000of10000000-40           4757960014     4957582268     +4.20%
BenchmarkSeries/10e6SeriesWithOneSample/10000of10000000-40           5287568989     4973603360     -5.94%
BenchmarkSeries/10e6SeriesWithOneSample/10000of10000000-40           4940028183     4996110747     +1.14%
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-40          4961982002     5150917278     +3.81%
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-40          4852168417     5274004293     +8.69%
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-40          4897943959     4966350836     +1.40%
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-40          4796559300     4966682721     +3.55%
BenchmarkSeries/10e6SeriesWithOneSample/100000of10000000-40          4870690517     5061585792     +3.92%
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-40         6044990597     6554478348     +8.43%
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-40         6848637524     6319292720     -7.73%
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-40         6270188862     6234622437     -0.57%
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-40         6630627177     6302021889     -4.96%
BenchmarkSeries/10e6SeriesWithOneSample/1000000of10000000-40         6886956538     6569022996     -4.62%
BenchmarkSeries/OneSeriesWith100e6Samples/1of100000000-40            37020405       37363983       +0.93%
BenchmarkSeries/OneSeriesWith100e6Samples/1of100000000-40            36836946       37665984       +2.25%
BenchmarkSeries/OneSeriesWith100e6Samples/1of100000000-40            36983663       37735010       +2.03%
BenchmarkSeries/OneSeriesWith100e6Samples/1of100000000-40            35769652       37063681       +3.62%
BenchmarkSeries/OneSeriesWith100e6Samples/1of100000000-40            36065700       37853475       +4.96%
BenchmarkSeries/OneSeriesWith100e6Samples/10of100000000-40           36591542       37292007       +1.91%
BenchmarkSeries/OneSeriesWith100e6Samples/10of100000000-40           38060806       37612714       -1.18%
BenchmarkSeries/OneSeriesWith100e6Samples/10of100000000-40           37189567       37127938       -0.17%
BenchmarkSeries/OneSeriesWith100e6Samples/10of100000000-40           34142731       37584568       +10.08%
BenchmarkSeries/OneSeriesWith100e6Samples/10of100000000-40           37778585       37958487       +0.48%
BenchmarkSeries/OneSeriesWith100e6Samples/100of100000000-40          36784314       36526013       -0.70%
BenchmarkSeries/OneSeriesWith100e6Samples/100of100000000-40          38017213       36485700       -4.03%
BenchmarkSeries/OneSeriesWith100e6Samples/100of100000000-40          34327648       36266622       +5.65%
BenchmarkSeries/OneSeriesWith100e6Samples/100of100000000-40          35331327       35490320       +0.45%
BenchmarkSeries/OneSeriesWith100e6Samples/100of100000000-40          36415368       34850967       -4.30%
BenchmarkSeries/OneSeriesWith100e6Samples/1000of100000000-40         36719457       36061775       -1.79%
BenchmarkSeries/OneSeriesWith100e6Samples/1000of100000000-40         36245498       36465980       +0.61%
BenchmarkSeries/OneSeriesWith100e6Samples/1000of100000000-40         36843385       36357582       -1.32%
BenchmarkSeries/OneSeriesWith100e6Samples/1000of100000000-40         36485731       36789888       +0.83%
BenchmarkSeries/OneSeriesWith100e6Samples/1000of100000000-40         34982941       37052282       +5.92%
BenchmarkSeries/OneSeriesWith100e6Samples/10000of100000000-40        38728394       37692706       -2.67%
BenchmarkSeries/OneSeriesWith100e6Samples/10000of100000000-40        38016653       36654963       -3.58%
BenchmarkSeries/OneSeriesWith100e6Samples/10000of100000000-40        37855045       38372409       +1.37%
BenchmarkSeries/OneSeriesWith100e6Samples/10000of100000000-40        37125852       37705730       +1.56%
BenchmarkSeries/OneSeriesWith100e6Samples/10000of100000000-40        36412315       37911913       +4.12%
BenchmarkSeries/OneSeriesWith100e6Samples/100000of100000000-40       37823423       37086613       -1.95%
BenchmarkSeries/OneSeriesWith100e6Samples/100000of100000000-40       37254053       38147762       +2.40%
BenchmarkSeries/OneSeriesWith100e6Samples/100000of100000000-40       38742010       37597268       -2.95%
BenchmarkSeries/OneSeriesWith100e6Samples/100000of100000000-40       36535632       38453071       +5.25%
BenchmarkSeries/OneSeriesWith100e6Samples/100000of100000000-40       37859349       38739158       +2.32%
BenchmarkSeries/OneSeriesWith100e6Samples/1000000of100000000-40      40913747       40044883       -2.12%
BenchmarkSeries/OneSeriesWith100e6Samples/1000000of100000000-40      39059499       40816649       +4.50%
BenchmarkSeries/OneSeriesWith100e6Samples/1000000of100000000-40      39375881       40097310       +1.83%
BenchmarkSeries/OneSeriesWith100e6Samples/1000000of100000000-40      40439518       39138862       -3.22%
BenchmarkSeries/OneSeriesWith100e6Samples/1000000of100000000-40      41235623       40420119       -1.98%
BenchmarkSeries/OneSeriesWith100e6Samples/10000000of100000000-40     127965441      129923216      +1.53%
BenchmarkSeries/OneSeriesWith100e6Samples/10000000of100000000-40     134440237      125991407      -6.28%
BenchmarkSeries/OneSeriesWith100e6Samples/10000000of100000000-40     123929699      133979820      +8.11%
BenchmarkSeries/OneSeriesWith100e6Samples/10000000of100000000-40     133013713      125939454      -5.32%
BenchmarkSeries/OneSeriesWith100e6Samples/10000000of100000000-40     128148652      125912643      -1.74%
```