@bwplotka 2020-04-19
Nice work!

The problem is really the benchmark of it.

The key part is marshalling and unmarshalling, so gRPC itself operations! 

With the benchmark we did.... we have ZERO of those operations, so you only measure splitting process part not the benefit of it (: 

So we have a few options:
* Create better micro benchmarks, one that involve actually some network traffic (not to test network latency, but rather marshalling /umarshalling HUGE responses). 
* Create better e2e benchmarks. So something on top of e2e framework or also https://github.com/thanos-io/thanosbench with some e2e tests
* Just don't benchmark here. Set an optional flag for splitting. THIS IS recommended for a reason by gRPC / protobuf guys so I assume there will be a benefit in real-life setups. With enabling this as optional tweak users can tweak it on production setups and see if it helps :hugs: 

I think we can do all those things.. But in this PR we can focus on third item, so flag, then focus on more benchmarking if you have time!