@bwplotka 2020-06-18
Just got pinged. I think the next think is to actually improve benchmark (: