@yeya24 2020-08-17
Don't think I have time to continue this now.